@@ -8,7 +8,7 @@
     "start:serve": "ng serve",
     "build": "ng build --prod",
     "test": "ng test",
-    "tests": "run-p test:*",
+    "tests": "yarn start:env && run-p  test:*",
     "test:dapp": "ng test dapp --browsers=ChromeHeadless --watch=false --progress=false",
     "test:services": "ng test services --browsers=ChromeHeadless --watch=false --progress=false",
     "test:ui": "ng test ui --browsers=ChromeHeadless --watch=false --progress=false",
@@ -22,16 +22,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.1.1",
-    "@angular/cdk": "^11.1.1",
-    "@angular/common": "~11.1.1",
-    "@angular/compiler": "~11.1.1",
-    "@angular/core": "~11.1.1",
-    "@angular/forms": "~11.1.1",
-    "@angular/material": "^11.1.1",
-    "@angular/platform-browser": "~11.1.1",
-    "@angular/platform-browser-dynamic": "~11.1.1",
-    "@angular/router": "~11.1.1",
+    "@angular/animations": "~11.2.3",
+    "@angular/cdk": "^11.2.2",
+    "@angular/common": "~11.2.3",
+    "@angular/compiler": "~11.2.3",
+    "@angular/core": "~11.2.3",
+    "@angular/forms": "~11.2.3",
+    "@angular/material": "^11.2.2",
+    "@angular/platform-browser": "~11.2.3",
+    "@angular/platform-browser-dynamic": "~11.2.3",
+    "@angular/router": "~11.2.3",
     "@ngneat/transloco": "^2.19.1",
     "@ngxd/core": "^10.0.0",
     "@types/lodash": "^4.14.168",
@@ -47,23 +47,22 @@
     "npm-run-all": "^4.1.5",
     "pm2": "^4.5.4",
     "prettier": "^2.2.1",
-    "rxjs": "~6.6.3",
-    "ts-standard": "^10.0.0",
+    "rxjs": "~6.6.6",
     "tslib": "^2.1.0",
     "tslint-etc": "^1.13.9",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1101.2",
+    "@angular-devkit/build-angular": "~0.1102.2",
     "@angular-eslint/builder": "1.0.0",
     "@angular-eslint/eslint-plugin": "1.0.0",
     "@angular-eslint/eslint-plugin-template": "1.0.0",
     "@angular-eslint/schematics": "1.0.0",
     "@angular-eslint/template-parser": "1.0.0",
-    "@angular/cli": "~11.1.2",
-    "@angular/compiler-cli": "~11.1.1",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular/cli": "~11.2.2",
+    "@angular/compiler-cli": "~11.2.3",
+    "@types/jasmine": "3.6.4",
+    "@types/jasminewd2": "2.0.8",
     "@types/node": "^12.11.1",
     "@typescript-eslint/eslint-plugin": "^4.14.1",
     "@typescript-eslint/parser": "4.3.0",
@@ -75,13 +74,13 @@
     "eslint-plugin-prefer-arrow": "1.2.2",
     "eslint-plugin-promise": "^4.2.1",
     "husky": "^4.3.8",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.2.3",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "jasmine-core": "3.6.0",
+    "jasmine-spec-reporter": "5.0.2",
+    "karma": "5.2.3",
+    "karma-chrome-launcher": "3.1.0",
+    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-jasmine": "4.0.1",
+    "karma-jasmine-html-reporter": "1.5.4",
     "lint-staged": "^10.4.2",
     "markdownlint-cli": "^0.24.0",
     "ng-packagr": "^11.1.2",
