@@ -1,38 +1,44 @@
-import { Inject, Injectable } from '@angular/core'
+import {Inject, Injectable} from '@angular/core'
 import {BehaviorSubject, combineLatest, Observable} from 'rxjs'
-import { filter, map, publishReplay, refCount } from 'rxjs/operators'
-import { ContractService } from '@services/contract/contract.service'
-import { UserService } from '@services/user/user.service'
-import { RolesInterface } from '@services/user/user.interface'
-import { GrantsVariationType, GrantTypesEnum } from './static.model'
+import {filter, map, publishReplay, refCount} from 'rxjs/operators'
+import {ContractService} from '@services/contract/contract.service'
+import {UserService} from '@services/user/user.service'
+import {RolesInterface} from '@services/user/user.interface'
+import {GrantsVariationType, GrantTypesEnum} from './static.model'
 import {translate, TranslocoService} from '@ngneat/transloco'
-import { API, AppApiInterface } from '@constants'
+import {API, AppApiInterface} from '@constants'
 
 @Injectable({
   providedIn: 'root'
 })
 export class StaticService {
-  public selectedContact$: BehaviorSubject<GrantTypesEnum> = new BehaviorSubject<GrantTypesEnum>( GrantTypesEnum.disruptive)
+  public selectedContact$: BehaviorSubject<GrantTypesEnum> = new BehaviorSubject<GrantTypesEnum>(GrantTypesEnum.disruptive)
 
   constructor (
     private readonly contractService: ContractService, // eslint-disable-line
     private readonly userService: UserService, // eslint-disable-line
     private readonly translocoService: TranslocoService, // eslint-disable-line
     @Inject(API) private readonly api: AppApiInterface // eslint-disable-line
-  ) { }
+  ) {
+  }
 
   public getContactsList (): Observable<GrantsVariationType[]> {
+
     const contracts = this.api.contracts as { [s: string]: string }
 
-    return this.translocoService.selectTranslateObject('contracts').pipe(
-      map((data: {[s: string]: GrantsVariationType}) => Object.keys(data).map((key) => ({
-        ...data[key],
-        name: key,
-        type: contracts[key] || null
-      } as GrantsVariationType))),
-      publishReplay(1),
-      refCount()
-    )
+    const res = this.translocoService.selectTranslateObject('contracts')
+      .pipe(
+        map((data: { [s: string]: GrantsVariationType }) => Object.keys(data).map((key) => ({
+          ...data[key],
+          name: key,
+          type: contracts[key] || null
+        } as GrantsVariationType))),
+        publishReplay(1),
+        refCount()
+      )
+    res.subscribe()
+
+    return res
   }
 
   getContactInfo (contactType: string): Observable<GrantsVariationType | null> {
