@@ -9,13 +9,13 @@ import {translate} from '@ngneat/transloco'
 import {DialogComponent} from '@ui/dialog/dialog.component'
 import {ApplyComponent} from '@ui/modals/apply/apply.component'
 import {
+  FinishApplicantsVotingArg,
   SubmitCallBackAcceptWorkResultArg,
   SubmitCallBackApplyArg,
   SubmitCallBackRewardArg
 } from '@ui/dialog/dialog.tokens'
 import {MatDialog} from '@angular/material/dialog'
 import {
-  TeamsControlsInterface,
   TemplateComponentAbstract,
   VoteTeamEventInterface
 } from '@pages/entity-page/entity.interface'
@@ -24,13 +24,18 @@ import {UserService} from '@services/user/user.service'
 import {AcceptWorkResultComponent} from '@ui/modals/accept-work-result/accept-work-result.component'
 import {combineLatest, Observable, Subject} from 'rxjs'
 import {
-  getWinnerTeamId, isAcceptWorkResultBtn,
-  isFinishApplicantsVoteBtn, isFinishVoteBtn, isShowAddRewardBtn, isStartWorkBtn,
-  teamsControls
+  getWinnerTeamId,
+  isAcceptWorkResultBtn,
+  isFinishApplicantsVoteBtn,
+  isFinishVoteBtn,
+  isShowAddRewardBtn,
+  isStartWorkBtn, prepareIsRejectBtnData,
+  prepareTeamsData,
+  prepareTeamsHeaderData
 } from '@pages/entity-page/disruptive-template/functions'
 import {ActivatedRoute} from '@angular/router'
 import {IScore} from '@services/interface'
-import {LinkHttpPipe} from '@libs/pipes/link-http.pipe'
+import {FinishApplicantsVotingComponent} from '@ui/modals/finish-applicants-voting/finish-applicants-voting.component'
 
 @Component({
   selector: 'app-disruptive-template',
@@ -45,7 +50,7 @@ export class DisruptiveTemplateComponent implements TemplateComponentAbstract, O
   grant$ = new Subject<ContractGrantModel>()
 
   private readonly destroyed$ = new Subject()
-  private linkHttpPipe: LinkHttpPipe = new LinkHttpPipe()
+
   private user$ = this.userService.data
     .pipe(
       takeUntil(this.destroyed$),
@@ -60,30 +65,11 @@ export class DisruptiveTemplateComponent implements TemplateComponentAbstract, O
       map((app: ContractGrantAppModel[]) => !!app.find((a) => !!a.process))
     )
 
-  public readonly teamsHeader$ = combineLatest(
+  public readonly teamsHeader$: Observable<IScore.IHeader> = combineLatest(
     [this.grant$, this.userService.data, this.userService.isBalanceMoreCommission$])
     .pipe(
       filter(([grant]) => grant !== null && grant !== undefined),
-      map(([grant, user, isBalance]): IScore.IHeader => {
-
-        let isProcess = false
-        grant?.app?.forEach( app => {
-          if(!!app.process?.value) {
-            isProcess = true
-          }
-        })
-
-        const res: IScore.IHeader = {
-          applyBtnText: translate('entity.apply'),
-          isApplyBtn: teamsControls(user, grant).isApplyBtn || false,
-          isBalanceMoreCommission: isBalance !== false,
-          isShowAppliers: grant?.isShowAppliers || false,
-          isUnauthorized: user.roles.isUnauthorized,
-          titleText: translate('entity.teams'),
-          isShowLogInForApplyBtn: user.roles.isUnauthorized && !isProcess
-        }
-        return res
-      })
+      map(([grant, user, isBalance]): IScore.IHeader => prepareTeamsHeaderData(grant, user, isBalance))
     )
 
   public readonly teams$: Observable<IScore.IUnit[]> = combineLatest(
@@ -92,111 +78,60 @@ export class DisruptiveTemplateComponent implements TemplateComponentAbstract, O
     .pipe(
       takeUntil(this.destroyed$),
       filter(([grant]) => grant !== null && grant !== undefined),
-      map(([grant, user, isProcess]) => {
-        const res: IScore.IUnit[] = []
-        const apps = grant.app || []
-        const controls = teamsControls(user, grant)
-        apps.forEach(app => {
-          const isCanVote =
-            (!(controls?.isVoteControls === 'hidden')&& this.grantStatusEnum.rejected !== grant?.status?.value)
-            && !controls?.voteFor.includes(app?.id?.value)
-          const score = app?.score?.value || 0
-          const unit: IScore.IUnit = {
-            isWinner: false,
-            isPerformer: !!app.process?.value,
-            isWinnerIcon: !!app.process?.value,
-            isPerformerIcon: true,
-            name: app.name.value,
-            solutionLink: null,
-            status: {
-              isSolution: false,
-              isRejected: false,
-              isApprove: false
-            },
-            square: {
-              score: app.score?.value || 0,
-              id: app?.id?.value || '',
-              isCanVote,
-              isShowResult: !user.roles.isDAO,
-            },
-            teamLink: this.linkHttpPipe.transform(app?.link?.value)
-          }
-          if(isProcess && score> 0 || !isProcess) {
-            res.push(unit)
-          }
-        })
-        return res
-      })
+      map(([grant, user, isProcess]) => prepareTeamsData(grant, user, isProcess))
     )
 
-  isShowStepperAndTeam$: Observable<boolean> = this.grant$
+  public readonly isShowStepperAndTeam$: Observable<boolean> = this.grant$
     .pipe(
       takeUntil(this.destroyed$),
       map(e => typeof e?.status?.value === 'string' ? e?.status?.value : ''),
       map(e => e !== GrantStatusEnum.rejected),
     )
 
-  isShowAddRewardBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
+  public readonly isShowAddRewardBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
     .pipe(
       takeUntil(this.destroyed$),
       map(([user, grant]) => isShowAddRewardBtn(user, grant))
     )
 
-  teamsControls$: Observable<TeamsControlsInterface> = combineLatest([this.userService.data, this.grant$])
-    .pipe(
-      takeUntil(this.destroyed$),
-      map(([user, grant]) => teamsControls(user, grant))
-    )
-
-  isStartWorkBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
+  public readonly isStartWorkBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
     .pipe(
       takeUntil(this.destroyed$),
       map(([user, grant]) => isStartWorkBtn(user, grant))
     )
 
-  isFinishApplicantsVoteBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
+  public readonly isFinishApplicantsVoteBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
     .pipe(
       takeUntil(this.destroyed$),
       map(([user, grant]) => isFinishApplicantsVoteBtn(user, grant))
     )
 
-  isAcceptWorkResultBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
+  public readonly isAcceptWorkResultBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
     .pipe(
       takeUntil(this.destroyed$),
       map(([user, grant]) => isAcceptWorkResultBtn(user, grant))
     )
 
-  isFinishVoteBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
+  public readonly isFinishVoteBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
     .pipe(
       takeUntil(this.destroyed$),
       map(([user, grant]) => isFinishVoteBtn(user, grant))
     )
 
-  voteForTaskData = {
+  public voteForTaskData = {
     isShow: false,
     isVote: false,
     isVoteInProcess: false
   }
 
-  isRejectBtn$ = combineLatest([this.userService.data, this.grant$])
+  public readonly isRejectBtn$: Observable<boolean> = combineLatest([this.userService.data, this.grant$])
     .pipe(
       takeUntil(this.destroyed$),
-      map(([user, grant]) => {
-        if (grant) {
-          const isWG = user.roles.isWG
-          const isStatusMatch = grant?.status?.value !== this.grantStatusEnum.rejected
-          return isWG && isStatusMatch
-        } else {
-          return false
-        }
-      })
+      map(([user, grant]) => prepareIsRejectBtnData(grant, user))
     )
 
 
-
-  @Input() set grant (data:
-                       ContractGrantModel
-  ) {
+  @Input() set grant (data: ContractGrantModel) {
     if (data !== this.inputGrant) {
       this.inputGrant = data
       this.prepareVoteForTaskData(data)
@@ -222,10 +157,7 @@ export class DisruptiveTemplateComponent implements TemplateComponentAbstract, O
   ) {
   }
 
-  vote (value:
-         'like' | 'dislike'
-  ):
-    void {
+  vote (value: 'like' | 'dislike'): void {
     const id = this.grant.id || ''
     this.voteForTaskData.isVoteInProcess = true
     this.disruptiveContractService.voteForTaskProposal(id, value).subscribe({
@@ -236,8 +168,7 @@ export class DisruptiveTemplateComponent implements TemplateComponentAbstract, O
     })
   }
 
-  signup ():
-    void {
+  signup (): void {
     this.signerService.login()
       .pipe(take(1))
       .subscribe(() => {
@@ -246,8 +177,7 @@ export class DisruptiveTemplateComponent implements TemplateComponentAbstract, O
       })
   }
 
-  openApplyModal ():
-    void {
+  openApplyModal (): void {
     const dialog = this.dialog.open(DialogComponent, {
       width: '500px',
       maxWidth: '100vw',
@@ -267,33 +197,26 @@ export class DisruptiveTemplateComponent implements TemplateComponentAbstract, O
     })
   }
 
-  voteTeam ($event:
-             VoteTeamEventInterface
-  ):
-    void {
+  voteTeam ($event: VoteTeamEventInterface): void {
     const id = this.grant?.id as string
     const teamId = $event.teamIdentifier
     const vote = $event.voteValue
     this.disruptiveContractService.voteForApplicant(id, teamId, vote).subscribe()
   }
 
-  finishVote ():
-    void {
+  finishVote (): void {
     this.disruptiveContractService.finishTaskProposalVoting(this.grant?.id as string).subscribe()
   }
 
-  startWork ():
-    void {
+  startWork (): void {
     this.disruptiveContractService.startWork(this.grant?.id as string).subscribe()
   }
 
-  reject ():
-    void {
+  reject (): void {
     this.disruptiveContractService.rejectTask(this.grant?.id as string).subscribe()
   }
 
-  acceptWorkResult ():
-    void {
+  acceptWorkResult (): void {
     const dialog = this.dialog.open(DialogComponent, {
       width: '500px',
       maxWidth: '100vw',
@@ -313,16 +236,41 @@ export class DisruptiveTemplateComponent implements TemplateComponentAbstract, O
     })
   }
 
-  finishApplicantsVote ():
-    void {
-    const id = this.grant.id as string
-    if (id) {
-      this.disruptiveContractService.finishApplicantsVoting(id, getWinnerTeamId(this.grant)).subscribe()
-    }
+  finishApplicantsVote (): void {
+
+    const teamIdList: string[] = []
+    this.grant.app?.forEach(el => {
+      if (el?.score?.value && +el?.score?.value > 0) {
+        teamIdList.push(el.id.value)
+      }
+    })
+
+    const dialog = this.dialog.open(DialogComponent, {
+      width: '500px',
+      maxWidth: '100vw',
+      data: {
+        component: FinishApplicantsVotingComponent,
+        params: {
+          title: translate('entity.finish_applicants_voting'),
+          submitBtnText: translate('modal.btn.propose_grant'),
+          grantId: this.grant?.id,
+          teamIdList,
+          proposedWinner: getWinnerTeamId(this.grant),
+          submitCallBack: (data: FinishApplicantsVotingArg) => {
+            const id = this.grant.id as string
+            const winnerTeamId = data.winnerTeamId
+            if (id && winnerTeamId) {
+              this.disruptiveContractService.finishApplicantsVoting(id, winnerTeamId).subscribe()
+            }
+            dialog.close()
+            this.cdr.markForCheck()
+          }
+        }
+      }
+    })
   }
 
-  addReward ():
-    void {
+  addReward (): void {
     const dialog = this.dialog.open(DialogComponent, {
       width: '500px',
       maxWidth: '100vw',
