@@ -2,6 +2,11 @@ import { UserDataInterface } from '@services/user/user.interface'
 import { ContractGrantModel } from '@services/contract/contract.model'
 import { GrantStatusEnum } from '@services/static/static.model'
 import { TeamsControlsInterface } from '@pages/entity-page/entity.interface'
+import {IScore} from '@services/interface'
+import {translate} from '@ngneat/transloco'
+import {LinkHttpPipe} from '@libs/pipes/link-http.pipe'
+
+const linkHttpPipe = new LinkHttpPipe()
 
 export const isFinishApplicantsVoteBtn = (user: UserDataInterface, grant: ContractGrantModel): boolean => {
   if (grant && grant.app) {
@@ -57,11 +62,13 @@ export const isAcceptWorkResultBtn = (user: UserDataInterface, grant: ContractGr
 }
 
 export const teamsControls = (user: UserDataInterface, grant: ContractGrantModel): TeamsControlsInterface => {
+
   const result: TeamsControlsInterface = {
     isVoteControls: 'show',
     voteFor: [],
     isApplyBtn: false
   }
+
   // isApplyBtn
   if (user.roles.isAuth && grant && grant.app && grant.status && grant.status.value) {
     const status = grant.status.value
@@ -74,20 +81,24 @@ export const teamsControls = (user: UserDataInterface, grant: ContractGrantModel
       }
     })
   }
+
   // isVoteControls
   if (grant.app) {
     grant.app.forEach((app) => {
-      if (app.key && app.voted && app.voted.value.includes(user.userAddress)) {
+      if (app.key && app.voted && app.voted.value && app.voted.value.includes(user.userAddress)) {
         result.voteFor.push(app.key)
       }
     })
   }
+
   if (grant && grant.status && grant.status.value && grant.status.value === GrantStatusEnum.readyToApply) {
     result.isVoteControls = 'show'
   } else {
     result.isVoteControls = 'hidden'
   }
+
   return result
+
 }
 
 export const isStartWorkBtn = (user: UserDataInterface, grant: ContractGrantModel): boolean => {
@@ -129,12 +140,115 @@ export const isShowAddRewardBtn = (user: UserDataInterface, grant: ContractGrant
 }
 
 export const isFinishVoteBtn = (user: UserDataInterface, grant: ContractGrantModel): boolean => {
+
   if (grant && grant.app) {
+
     const isVoted = !!grant.voted
     const isStatusMatch = grant?.status?.value === GrantStatusEnum.proposed
     const isWG = user.roles.isWG
+
     return isVoted && isWG && isStatusMatch
+
   } else {
+
     return false
+
   }
+
 }
+
+export const prepareTeamsHeaderData = (
+  grant: ContractGrantModel,
+  user: UserDataInterface,
+  isBalance: boolean
+): IScore.IHeader => {
+
+  let isProcess = false
+  grant?.app?.forEach( app => {
+    if(!!app.process?.value) {
+      isProcess = true
+    }
+  })
+
+  const res: IScore.IHeader = {
+    applyBtnText: translate('entity.apply'),
+    isApplyBtn: teamsControls(user, grant).isApplyBtn || false,
+    isBalanceMoreCommission: isBalance !== false,
+    isShowAppliers: grant?.isShowAppliers || false,
+    isUnauthorized: user.roles.isUnauthorized,
+    titleText: translate('entity.teams'),
+    isShowLogInForApplyBtn: user.roles.isUnauthorized && !isProcess
+  }
+
+  return res
+
+}
+
+export const prepareTeamsData = (
+  grant: ContractGrantModel,
+  user: UserDataInterface,
+  isProcess: boolean
+): IScore.IUnit[] => {
+
+  const res: IScore.IUnit[] = []
+  const apps = grant.app || []
+  const controls = teamsControls(user, grant)
+
+  apps.forEach(app => {
+
+    const isCanVote =
+      (!(controls?.isVoteControls === 'hidden')&& GrantStatusEnum.rejected !== grant?.status?.value)
+      && !controls?.voteFor.includes(app?.id?.value)
+    const score = app?.score?.value || 0
+
+    const unit: IScore.IUnit = {
+      isWinner: false,
+      isPerformer: !!app.process?.value,
+      isWinnerIcon: !!app.process?.value,
+      isPerformerIcon: true,
+      name: app.name.value,
+      solutionLink: null,
+      status: {
+        isSolution: false,
+        isRejected: false,
+        isApprove: false
+      },
+      square: {
+        score: app.score?.value || 0,
+        id: app?.id?.value || '',
+        isCanVote,
+        isShowResult: !user.roles.isDAO,
+      },
+      teamLink: linkHttpPipe.transform(app?.link?.value)
+    }
+
+    if(isProcess && score> 0 || !isProcess) {
+      res.push(unit)
+    }
+
+  })
+
+  return res
+
+}
+
+export const prepareIsRejectBtnData = (
+  grant: ContractGrantModel,
+  user: UserDataInterface,
+): boolean => {
+
+  if (grant) {
+
+    const isWG = user.roles.isWG
+    const isStatusMatch = grant?.status?.value !== GrantStatusEnum.rejected
+
+    return isWG && isStatusMatch
+
+  } else {
+
+    return false
+
+  }
+
+}
+
